+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Removed - 删除功能/接口
 Fixed - 修复问题
 Others - 其他 
 -->
+## 2025-11-6 v0.1.31 【普通更新】
+### Fixed
+* 普通：修复glx模式下 createVideo underGameView为true 黑屏的问题
+* 普通：修复build profile在团结1.6之后的导出问题
+
 ## 2025-11-6 v0.1.30 【重要更新】
 ### Feature
 * 重要：EmscriptenGLX支持微信压缩纹理